+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Ansible Changes By Release
 * make remote_md5 internal function work with non-bash shells
 * allow user to be passed in via --extra-vars (regression)
 * add ability to store the result of any command in a register (see examples/playbooks/register_logic.yml)
+* add --limit option, which can be used to further confine the pattern given in ansible-playbooks
 
 0.6 "Cabo" -- August 6, 2012